Finalize jax.numpy array API compliance
