+++ jax/_src/basearray.pyi
@@ -213,6 +213,13 @@ class Array(abc.ABC):
   def traceback(self) -> Traceback: ...
   def unsafe_buffer_pointer(self) -> int: ...
 
+  # Array API inclusions
+  @property
+  def device(self) -> Sharding: ...
+  def to_device(self, device: Device | Sharding | None, *,
+               stream: int | Any | None = None) -> Sharding: ...
+  def __array_namespace__(self, /, *,
+                          api_version: None | str = None): ...
 
 ArrayLike = Union[
   Array,  # JAX array type
